Include changes from original project  (#4)
* lottiemodel: Improve opacity calculation for gradient stop

Basically, Graeidnt stop's color and opacity are provided as separate arrays.
Stop position and opacity position do not match each other.
Existing code is a sequential approach. It caused problems in various cases of positions.
The improved logic repeats the loop, but no exceptions are raised. It's not complicated, it's simple.

This code referenced the lottie-android library.
https://github.com/airbnb/lottie-android/blob/master/lottie/src/main/java/com/airbnb/lottie/parser/GradientColorParser.java

* replace uint ushort and uchar typedefs with uint32 uint16 and uint8
-- added format script to run clang format on latest commit.

* replace unsigned int with uint32_t in Int()

Signed-off-by: wangxuedong <wangxuedong@xiaomi.com>
Change-Id: I4390c12fcf3aefccfe20290436b0dad96877008c

* Add lottie_init() and lottie_shutdown() c api.

To support dynamic loading and unloading of rlottie library safely
we need to deallocate the resource cache as well as safely shutdown all the
worker threads.
current patch only stops the Render and Rle task schedulers when lottie_shutdown is called.

Things yet to be implemented during shutdown phase
- Unload image loader if loaded dynamically.
- Check if we can release some cache resources.
- Currently multiple load and unload of rlottie library will not work as we are not starting the
  scheduler again when lottie_init() called multiple time in the same process.

* add formal parameter void to lottie_init() and lottie_shutdown()

Signed-off-by: wangxuedong <wangxuedong@xiaomi.com>
Change-Id: I40dbe4b07f9581e5042e468d3860e135622580c8

* Ignore animations with objects of unspecified type

* Check border of color table while generating gradient

* Delete accidentally committed Vim swap file

* Fix one frame missing (#529)

As mStartFrame and mEndFrame was counted from 0 and totalFrame() was
calculated as a difference, there were always one frame missing and
the animation rescaled. This patch adds one to the total frames count.

issue: #527

* example: lottieviewer - fixed frame no (#528)

There is a problem (issue #527) with totalFrame() in rlottie.
As lottieviewer didn't showed the actual frame number, but calculated it from
progress, it camouflaged the problem. Now the actual frame number and the
total number of frames are displayed.

Signed-off-by: wangxuedong <wangxuedong@xiaomi.com>
Co-authored-by: JunsuChoi <jsuya.choi@samsung.com>
Co-authored-by: Subhransu Mohanty <sub.mohanty@samsung.com>
Co-authored-by: wangxuedong <wangxuedong@xiaomi.com>
Co-authored-by: Nicholas Guriev <nicholas@guriev.su>
Co-authored-by: Michal Maciola <71131832+mmaciola@users.noreply.github.com>